Move core::collector::skb mod.rs -> skb.rs
