+++ src/collector/skb/mod.rs
@@ -4,27 +4,7 @@
 //! `struct sk_buff *` as a parameter), to filter skbs, and to track them;
 //! allowing to reconstruct their path in the Linux networking stack.
 
-use anyhow::Result;
-
-use super::Collector;
-use crate::core::probe::kernel;
-
-pub(super) struct SkbCollector {}
-
-impl Collector for SkbCollector {
-    fn new() -> Result<SkbCollector> {
-        Ok(SkbCollector {})
-    }
-
-    fn name(&self) -> &'static str {
-        "skb"
-    }
-
-    fn init(&mut self, _kernel: &mut kernel::Kernel) -> Result<()> {
-        Ok(())
-    }
-
-    fn start(&mut self) -> Result<()> {
-        Ok(())
-    }
-}
+// Re-export skb.rs
+#[allow(clippy::module_inception)]
+pub(super) mod skb;
+pub(super) use skb::*;
